#27 - refactor operation files path
