+++ feature/operations/src/androidTest/java/br/com/rbrthmn/operations/DatePickerFieldUITest.kt
@@ -27,8 +27,8 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
-import br.com.rbrthmn.operations.components.CALENDAR_TEST_TAG
-import br.com.rbrthmn.operations.components.DatePickerField
+import br.com.rbrthmn.operations.ui.components.CALENDAR_TEST_TAG
+import br.com.rbrthmn.operations.ui.components.DatePickerField
 import br.com.rbrthmn.ui.BaseUITest
 import br.com.rbrthmn.ui.onNodeWithStringId
 import org.junit.Test
