+++ feature/operations/src/main/java/br/com/rbrthmn/operations/ui/OperationsScreenUiState.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.operations
+package br.com.rbrthmn.operations.ui
 
 import androidx.compose.runtime.Composable
 import java.time.LocalDate
