+++ feature/operations/src/main/java/br/com/rbrthmn/operations/ui/components/OperationsListCard.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.operations.components
+package br.com.rbrthmn.operations.ui.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -64,10 +64,10 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
-import br.com.rbrthmn.operations.Operation
-import br.com.rbrthmn.operations.OperationType
-import br.com.rbrthmn.operations.OperationsScreenContract
-import br.com.rbrthmn.operations.OperationsScreenViewModel
+import br.com.rbrthmn.operations.ui.Operation
+import br.com.rbrthmn.operations.ui.OperationType
+import br.com.rbrthmn.operations.ui.OperationsScreenContract
+import br.com.rbrthmn.operations.ui.OperationsScreenViewModel
 import br.com.rbrthmn.operations.R
 import br.com.rbrthmn.ui.utils.ResourceStringProvider
 import br.com.rbrthmn.ui.utils.valueWithCurrencyString
